all: join app error types in a single file
