+++ types/app/app.go
@@ -4,8 +4,6 @@
 
 package app
 
-import "errors"
-
 type AppRouter struct {
 	Name         string            `json:"name"`
 	Opts         map[string]string `json:"opts"`
@@ -14,7 +12,3 @@ type AppRouter struct {
 	Status       string            `json:"status,omitempty" bson:"-"`
 	StatusDetail string            `json:"status-detail,omitempty" bson:"-"`
 }
-
-var (
-	ErrAppNotFound = errors.New("App not found")
-)
